+++ nautobot/core/tests/test_views.py
@@ -1,11 +1,14 @@
 import re
 import urllib.parse
 
+from django.contrib.contenttypes.models import ContentType
 from django.test import override_settings
 from django.test.utils import override_script_prefix
 from django.urls import get_script_prefix, reverse
 
-from nautobot.utilities.testing import TestCase
+from nautobot.core.models.dynamic_groups import DynamicGroup
+from nautobot.dcim.models import Device
+from nautobot.utilities.testing import TestCase, ViewTestCases
 
 
 class HomeViewTestCase(TestCase):
@@ -268,3 +271,40 @@ class LoginUI(TestCase):
         self.client.logout()
         sso_login_search_result = self.make_request()
         self.assertIsNotNone(sso_login_search_result)
+
+
+class DynamicGroupTestCase(
+    ViewTestCases.CreateObjectViewTestCase,
+    ViewTestCases.DeleteObjectViewTestCase,
+    ViewTestCases.EditObjectViewTestCase,
+    ViewTestCases.GetObjectViewTestCase,
+    ViewTestCases.GetObjectChangelogViewTestCase,
+    ViewTestCases.ListObjectsViewTestCase,
+    ViewTestCases.BulkDeleteObjectsViewTestCase,
+    # NOTE: This isn't using `ViewTestCases.PrimaryObjectViewTestCase` because bulk-import/edit
+    # views for DynamicGroup do not make sense at this time, primarily because `content_type` is
+    # immutable after create.
+):
+    model = DynamicGroup
+
+    @classmethod
+    def setUpTestData(cls):
+
+        content_type = ContentType.objects.get_for_model(Device)
+
+        # DynamicGroup objects to test.
+        DynamicGroup.objects.create(name="DG 1", slug="dg-1", content_type=content_type)
+        DynamicGroup.objects.create(name="DG 2", slug="dg-2", content_type=content_type)
+        DynamicGroup.objects.create(name="DG 3", slug="dg-3", content_type=content_type)
+
+        cls.form_data = {
+            "name": "new_dynamic_group",
+            "slug": "new-dynamic-group",
+            "description": "I am a new dynamic group object.",
+            "content_type": content_type.pk,
+            # Management form fields required for the dynamic formset
+            "dynamic_group_memberships-TOTAL_FORMS": "0",
+            "dynamic_group_memberships-INITIAL_FORMS": "1",
+            "dynamic_group_memberships-MIN_NUM_FORMS": "0",
+            "dynamic_group_memberships-MAX_NUM_FORMS": "1000",
+        }
